mask_blending/laplacian_blend/requirements.txt
numpy>=1.21.0
Pillow>=8.0.0
opencv-python>=4.5.0
matplotlib>=3.3.0
jupyter>=1.0.0 